+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/arg_parser.go
@@ -9,6 +9,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_errors"
+	"github.com/kurtosis-tech/kurtosis/core/server/commons/enclave_data_directory"
 	"go.starlark.net/starlark"
 	"go.starlark.net/starlarkstruct"
 	"math"
@@ -18,16 +19,16 @@ import (
 
 const (
 	serviceIdArgName     = "service_id"
-	serviceConfigArgName = "service_config"
+	serviceConfigArgName = "config"
 	defineFactArgName    = "define_fact"
 
-	containerImageNameKey          = "container_image_name"
+	containerImageNameKey          = "image"
 	factNameArgName                = "fact_name"
-	usedPortsKey                   = "used_ports"
+	usedPortsKey                   = "ports"
 	entryPointArgsKey              = "entry_point_args"
 	cmdArgsKey                     = "cmd_args"
 	envVarArgsKey                  = "env_vars"
-	filesArtifactMountDirpathsKey  = "files_artifact_mount_dirpaths"
+	filesArtifactMountDirpathsKey  = "files"
 	portIdKey                      = "port_id"
 	requestEndpointKey             = "endpoint"
 	requestMethodEndpointKey       = "method"
@@ -184,6 +185,17 @@ func ParseFilePath(filePathArgName string, filePathStr starlark.String) (string,
 	return srcPath, nil
 }
 
+func ParseArtifactUuid(artifactUuidArgName string, artifactUuidStr starlark.String) (enclave_data_directory.FilesArtifactUUID, *startosis_errors.InterpretationError) {
+	artifactUuid, interpretationErr := safeCastToString(artifactUuidStr, artifactUuidArgName)
+	if interpretationErr != nil {
+		return "", interpretationErr
+	}
+	if len(artifactUuid) == 0 {
+		return "", startosis_errors.NewInterpretationError("Artifact Uuid can't be empty for argument '%s'", artifactUuidArgName)
+	}
+	return enclave_data_directory.FilesArtifactUUID(artifactUuid), interpretationErr
+}
+
 func ParseTemplatesAndData(templatesAndData *starlark.Dict) (map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData, *startosis_errors.InterpretationError) {
 	templateAndDataByDestRelFilepath := make(map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData)
 	for _, relPathInFilesArtifactKey := range templatesAndData.Keys() {
@@ -249,7 +261,8 @@ func parseServiceConfigContainerImageName(serviceConfig *starlarkstruct.Struct) 
 func parseServiceConfigPrivatePorts(serviceConfig *starlarkstruct.Struct) (map[string]*kurtosis_core_rpc_api_bindings.Port, *startosis_errors.InterpretationError) {
 	privatePortsRawArg, err := serviceConfig.Attr(usedPortsKey)
 	if err != nil {
-		return nil, startosis_errors.NewInterpretationError("Missing `%s` as part of the service config", usedPortsKey)
+		// not all services need to create ports, this being empty is okay
+		return map[string]*kurtosis_core_rpc_api_bindings.Port{}, nil
 	}
 	privatePortsArg, ok := privatePortsRawArg.(*starlark.Dict)
 	if !ok {