+++ core/server/api_container/server/startosis_engine/startosis_interpreter_test.go
@@ -18,6 +18,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/kurtosis_instruction/upload_files"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_errors"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_modules/mock_module_content_provider"
+	"github.com/kurtosis-tech/kurtosis/core/server/commons/enclave_data_directory"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.starlark.net/starlark"
@@ -142,14 +143,14 @@ print("Starting Startosis script!")
 service_id = "example-datastore-server"
 print("Adding service " + service_id)
 
-service_config = struct(
-	container_image_name = "` + testContainerImageName + `",
-	used_ports = {
+config = struct(
+	image = "` + testContainerImageName + `",
+	ports = {
 		"grpc": struct(number = 1323, protocol = "TCP")
 	},
 	private_ip_address_placeholder = "` + privateIPAddressPlaceholder + `"
 )
-datastore_service = add_service(service_id = service_id, service_config = service_config)
+datastore_service = add_service(service_id = service_id, config = config)
 print("The grpc port is " + str(datastore_service.ports["grpc"].number))
 print("The grpc port protocol is " + datastore_service.ports["grpc"].protocol)
 print("The datastore service ip address is " + datastore_service.ip_address)
@@ -181,13 +182,13 @@ print("Starting Startosis script!")
 service_id = "example-datastore-server"
 print("Adding service " + service_id)
 
-service_config = struct(
+config = struct(
 	# /!\ /!\ missing container name /!\ /!\
-	used_ports = {
+	ports = {
 		"grpc": struct(number = 1323, protocol = "TCP")
 	}
 )
-add_service(service_id = service_id, service_config = service_config)
+add_service(service_id = service_id, config = config)
 `
 
 	scriptOutput, interpretationError, instructions := interpreter.Interpret(context.Background(), ModuleIdPlaceholderForStandaloneScripts, script, EmptyInputArgs)
@@ -199,7 +200,7 @@ add_service(service_id = service_id, service_config = service_config)
 			*startosis_errors.NewCallFrame("<toplevel>", startosis_errors.NewScriptPosition(13, 12)),
 			*startosis_errors.NewCallFrame("add_service", startosis_errors.NewScriptPosition(0, 0)),
 		},
-		"Evaluation error: Missing value 'container_image_name' as element of the struct object 'service_config'",
+		"Evaluation error: Missing value 'image' as element of the struct object 'config'",
 	)
 	require.Equal(t, expectedError, interpretationError)
 }
@@ -214,13 +215,13 @@ print("Starting Startosis script!")
 service_id = "example-datastore-server"
 print("Adding service " + service_id)
 
-service_config = struct(
-	container_image_name = "` + testContainerImageName + `",
-	used_ports = {
+config = struct(
+	image = "` + testContainerImageName + `",
+	ports = {
 		"grpc": struct(number = 1323, protocol = "TCPK") # typo in protocol
 	}
 )
-add_service(service_id = service_id, service_config = service_config)
+add_service(service_id = service_id, config = config)
 `
 
 	scriptOutput, interpretationError, instructions := interpreter.Interpret(context.Background(), ModuleIdPlaceholderForStandaloneScripts, script, EmptyInputArgs)
@@ -246,13 +247,13 @@ print("Starting Startosis script!")
 service_id = "example-datastore-server"
 print("Adding service " + service_id)
 
-service_config = struct(
-	container_image_name = "` + testContainerImageName + `",
-	used_ports = {
+config = struct(
+	image = "` + testContainerImageName + `",
+	ports = {
 		"grpc": struct(number = "1234", protocol = "TCP") # port number should be an int
 	}
 )
-add_service(service_id = service_id, service_config = service_config)
+add_service(service_id = service_id, config = config)
 `
 
 	scriptOutput, interpretationError, instructions := interpreter.Interpret(context.Background(), ModuleIdPlaceholderForStandaloneScripts, script, EmptyInputArgs)
@@ -282,16 +283,16 @@ def deploy_datastore_services():
     for i in range(len(ports)):
         unique_service_id = service_id + "-" + str(i)
         print("Adding service " + unique_service_id)
-        service_config = struct(
-			container_image_name = "` + testContainerImageName + `",
-			used_ports = {
+        config = struct(
+			image = "` + testContainerImageName + `",
+			ports = {
 				"grpc": struct(
 					number = ports[i],
 					protocol = "TCP"
 				)
 			}
 		)
-        add_service(service_id = unique_service_id, service_config = service_config)
+        add_service(service_id = unique_service_id, config = config)
 
 deploy_datastore_services()
 print("Done!")
@@ -452,10 +453,10 @@ func TestStartosisInterpreter_ValidSimpleScriptWithImportedStruct(t *testing.T) 
 	moduleBar := "github.com/foo/bar/lib.star"
 	seedModules[moduleBar] = `
 service_id = "example-datastore-server"
-print("Constructing service_config")
-service_config = struct(
-	container_image_name = "kurtosistech/example-datastore-server",
-	used_ports = {
+print("Constructing config")
+config = struct(
+	image = "kurtosistech/example-datastore-server",
+	ports = {
 		"grpc": struct(number = 1323, protocol = "TCP")
 	}
 )
@@ -465,11 +466,11 @@ service_config = struct(
 	require.Nil(t, moduleContentProvider.BulkAddFileContent(seedModules))
 	interpreter := NewStartosisInterpreter(testServiceNetwork, moduleContentProvider)
 	script := `
-load("` + moduleBar + `", "service_id", "service_config")
+load("` + moduleBar + `", "service_id", "config")
 print("Starting Startosis script!")
 
 print("Adding service " + service_id)
-add_service(service_id = service_id, service_config = service_config)
+add_service(service_id = service_id, config = config)
 `
 
 	scriptOutput, interpretationError, instructions := interpreter.Interpret(context.Background(), ModuleIdPlaceholderForStandaloneScripts, script, EmptyInputArgs)
@@ -480,7 +481,7 @@ add_service(service_id = service_id, service_config = service_config)
 
 	require.Equal(t, instructions[0], addServiceInstruction)
 
-	expectedOutput := `Constructing service_config
+	expectedOutput := `Constructing config
 Starting Startosis script!
 Adding service example-datastore-server
 `
@@ -498,16 +499,16 @@ def deploy_datastore_services():
     for i in range(len(ports)):
         unique_service_id = service_id + "-" + str(i)
         print("Adding service " + unique_service_id)
-        service_config = struct(
-			container_image_name = "kurtosistech/example-datastore-server",
-			used_ports = {
+        config = struct(
+			image = "kurtosistech/example-datastore-server",
+			ports = {
 				"grpc": struct(
 					number = ports[i],
 					protocol = "TCP"
 				)
 			}
 		)
-        add_service(service_id = unique_service_id, service_config = service_config)
+        add_service(service_id = unique_service_id, config = config)
 `
 	moduleContentProvider := mock_module_content_provider.NewMockModuleContentProvider()
 	defer moduleContentProvider.RemoveAll()
@@ -547,22 +548,22 @@ func TestStartosisInterpreter_AddServiceInOtherModulePopulatesQueue(t *testing.T
 	moduleBar := "github.com/foo/bar/lib.star"
 	seedModules[moduleBar] = `
 service_id = "example-datastore-server"
-print("Constructing service_config")
-service_config = struct(
-	container_image_name = "kurtosistech/example-datastore-server",
-	used_ports = {
+print("Constructing config")
+config = struct(
+	image = "kurtosistech/example-datastore-server",
+	ports = {
 		"grpc": struct(number = 1323, protocol = "TCP")
 	}
 )
 print("Adding service " + service_id)
-add_service(service_id = service_id, service_config = service_config)
+add_service(service_id = service_id, config = config)
 `
 	moduleContentProvider := mock_module_content_provider.NewMockModuleContentProvider()
 	defer moduleContentProvider.RemoveAll()
 	require.Nil(t, moduleContentProvider.BulkAddFileContent(seedModules))
 	interpreter := NewStartosisInterpreter(testServiceNetwork, moduleContentProvider)
 	script := `
-load("` + moduleBar + `", "service_id", "service_config")
+load("` + moduleBar + `", "service_id", "config")
 print("Starting Startosis script!")
 `
 
@@ -574,7 +575,7 @@ print("Starting Startosis script!")
 
 	require.Equal(t, instructions[0], addServiceInstruction)
 
-	expectedOutput := `Constructing service_config
+	expectedOutput := `Constructing config
 Adding service example-datastore-server
 Starting Startosis script!
 `
@@ -586,27 +587,27 @@ func TestStartosisInterpreter_TestInstructionQueueAndOutputBufferDontHaveDupesIn
 	moduleBar := "github.com/foo/bar/lib.star"
 	seedModules[moduleBar] = `
 service_id = "example-datastore-server"
-print("Constructing service_config")
-service_config = struct(
-	container_image_name = "kurtosistech/example-datastore-server",
-	used_ports = {
+print("Constructing config")
+config = struct(
+	image = "kurtosistech/example-datastore-server",
+	ports = {
 		"grpc": struct(number = 1323, protocol = "TCP")
 	}
 )
 print("Adding service " + service_id)
-add_service(service_id = service_id, service_config = service_config)
+add_service(service_id = service_id, config = config)
 `
 	moduleContentProvider := mock_module_content_provider.NewMockModuleContentProvider()
 	defer moduleContentProvider.RemoveAll()
 	require.Nil(t, moduleContentProvider.BulkAddFileContent(seedModules))
 	interpreter := NewStartosisInterpreter(testServiceNetwork, moduleContentProvider)
 	scriptA := `
-load("` + moduleBar + `", "service_id", "service_config")
+load("` + moduleBar + `", "service_id", "config")
 print("Starting Startosis script!")
 `
 	addServiceInstructionFromScriptA := createSimpleAddServiceInstruction(t, "example-datastore-server", testContainerImageName, 1323, 11, 12, moduleBar, defaultEntryPointArgs, defaultCmdArgs, defaultEnvVars, defaultPrivateIPAddressPlaceholder)
 
-	expectedOutputFromScriptA := `Constructing service_config
+	expectedOutputFromScriptA := `Constructing config
 Adding service example-datastore-server
 Starting Startosis script!
 `
@@ -623,13 +624,13 @@ print("Starting Startosis script!")
 service_id = "example-datastore-server"
 print("Adding service " + service_id)
 
-service_config = struct(
-	container_image_name = "kurtosistech/example-datastore-server",
-	used_ports = {
+config = struct(
+	image = "kurtosistech/example-datastore-server",
+	ports = {
 		"grpc": struct(number = 1323, protocol = "TCP")
 	}
 )
-add_service(service_id = service_id, service_config = service_config)
+add_service(service_id = service_id, config = config)
 `
 	addServiceInstructionFromScriptB := createSimpleAddServiceInstruction(t, "example-datastore-server", testContainerImageName, 1323, 13, 12, starlarkFilenamePlaceholderAsNotUsed, defaultEntryPointArgs, defaultCmdArgs, defaultEnvVars, defaultPrivateIPAddressPlaceholder)
 	expectedOutputFromScriptB := `Starting Startosis script!
@@ -651,25 +652,25 @@ func TestStartosisInterpreter_AddServiceWithEnvVarsCmdArgsAndEntryPointArgs(t *t
 print("Starting Startosis script!")
 service_id = "example-datastore-server"
 print("Adding service " + service_id)
-store_service_config = struct(
-	container_image_name = "kurtosistech/example-datastore-server",
-	used_ports = {
+store_config = struct(
+	image = "kurtosistech/example-datastore-server",
+	ports = {
 		"grpc": struct(number = 1323, protocol = "TCP")
 	}
 )
-datastore_service = add_service(service_id = service_id, service_config = store_service_config)
+datastore_service = add_service(service_id = service_id, config = store_config)
 client_service_id = "example-datastore-client"
 print("Adding service " + client_service_id)
-client_service_config = struct(
-	container_image_name = "kurtosistech/example-datastore-client",
-	used_ports = {
+client_config = struct(
+	image = "kurtosistech/example-datastore-client",
+	ports = {
 		"grpc": struct(number = 1337, protocol = "TCP")
 	},
 	entry_point_args = ["--store-port " + str(datastore_service.ports["grpc"].number), "--store-ip " + datastore_service.ip_address],
 	cmd_args = ["ping", datastore_service.ip_address],
 	env_vars = {"STORE_IP": datastore_service.ip_address}
 )
-add_service(service_id = client_service_id, service_config = client_service_config)
+add_service(service_id = client_service_id, config = client_config)
 `
 
 	scriptOutput, interpretationError, instructions := interpreter.Interpret(context.Background(), ModuleIdPlaceholderForStandaloneScripts, script, EmptyInputArgs)
@@ -750,12 +751,14 @@ exec(service_id = "example-datastore-server", command = ["mkdir", "/tmp/foo"], e
 }
 
 func TestStartosisInterpreter_StoreFileFromService(t *testing.T) {
+	testArtifactUuid, err := enclave_data_directory.NewFilesArtifactUUID()
+	require.Nil(t, err)
 	moduleContentProvider := mock_module_content_provider.NewMockModuleContentProvider()
 	defer moduleContentProvider.RemoveAll()
 	interpreter := NewStartosisInterpreter(testServiceNetwork, moduleContentProvider)
 	script := `
 print("Storing file from service!")
-artifact_uuid=store_file_from_service(service_id="example-datastore-server", src_path="/foo/bar")
+artifact_uuid=store_file_from_service(service_id="example-datastore-server", src_path="/foo/bar", artifact_uuid="` + string(testArtifactUuid) + `")
 print(artifact_uuid)
 `
 
@@ -768,14 +771,15 @@ print(artifact_uuid)
 		*kurtosis_instruction.NewInstructionPosition(3, 38, starlarkFilenamePlaceholderAsNotUsed),
 		"example-datastore-server",
 		"/foo/bar",
+		testArtifactUuid,
 	)
 
 	require.Equal(t, instructions[0], storeInstruction)
 
 	expectedOutput := fmt.Sprintf(`Storing file from service!
-{{kurtosis:%v-3:38.artifact_uuid}}
-`, starlarkFilenamePlaceholderAsNotUsed)
-	require.Equal(t, expectedOutput, string(scriptOutput))
+%v
+`, testArtifactUuid)
+	require.Regexp(t, expectedOutput, string(scriptOutput))
 }
 
 func TestStartosisInterpreter_ReadFileFromGithub(t *testing.T) {
@@ -821,6 +825,8 @@ wait(service_id="%v", fact_name="%v")
 }
 
 func TestStartosisInterpreter_RenderTemplates(t *testing.T) {
+	testArtifactUuid, err := enclave_data_directory.NewFilesArtifactUUID()
+	require.Nil(t, err)
 	moduleContentProvider := mock_module_content_provider.NewMockModuleContentProvider()
 	defer moduleContentProvider.RemoveAll()
 	interpreter := NewStartosisInterpreter(testServiceNetwork, moduleContentProvider)
@@ -841,7 +847,7 @@ data = {
 		"template_data_json": encoded_json
     }
 }
-artifact_uuid = render_templates(template_and_data_by_dest_rel_filepath = data)
+artifact_uuid = render_templates(template_and_data_by_dest_rel_filepath = data, artifact_uuid = "` + string(testArtifactUuid) + `")
 print(artifact_uuid)
 `
 
@@ -873,14 +879,16 @@ print(artifact_uuid)
 		templateAndDataByDestFilepath,
 		starlark.StringDict{
 			"template_and_data_by_dest_rel_filepath": templateAndDataValues,
+			"artifact_uuid":                          starlark.String(testArtifactUuid),
 		},
+		testArtifactUuid,
 	)
 
 	require.Equal(t, renderInstruction, instructions[0])
 
 	expectedOutput := fmt.Sprintf(`Rendering template to disk!
-{{kurtosis:%v-18:33.artifact_uuid}}
-`, starlarkFilenamePlaceholderAsNotUsed)
+%v
+`, testArtifactUuid)
 	require.Equal(t, expectedOutput, string(scriptOutput))
 }
 
@@ -1102,11 +1110,13 @@ proto: (line 1:15): invalid value for string type: 3`)
 }
 
 func TestStartosisInterpreter_ThreeLevelNestedInstructionPositionTest(t *testing.T) {
+	testArtifactUuid, err := enclave_data_directory.NewFilesArtifactUUID()
+	require.Nil(t, err)
 	storeFileDefinitionPath := "github.com/kurtosis/store.star"
 	storeFileContent := `
 def store_for_me():
 	print("In the store files instruction")
-	artifact_uuid=store_file_from_service(service_id="example-datastore-server", src_path="/foo/bar")
+	artifact_uuid=store_file_from_service(service_id="example-datastore-server", src_path="/foo/bar", artifact_uuid = "` + string(testArtifactUuid) + `")
 	return artifact_uuid
 `
 
@@ -1120,7 +1130,7 @@ def call_store_for_me():
 
 	moduleContentProvider := mock_module_content_provider.NewMockModuleContentProvider()
 	defer moduleContentProvider.RemoveAll()
-	err := moduleContentProvider.AddFileContent(storeFileDefinitionPath, storeFileContent)
+	err = moduleContentProvider.AddFileContent(storeFileDefinitionPath, storeFileContent)
 	require.Nil(t, err)
 
 	err = moduleContentProvider.AddFileContent(moduleThatCallsStoreFile, moduleThatCallsStoreFileContent)
@@ -1142,14 +1152,15 @@ print(uuid)
 		*kurtosis_instruction.NewInstructionPosition(4, 39, storeFileDefinitionPath),
 		"example-datastore-server",
 		"/foo/bar",
+		testArtifactUuid,
 	)
 
 	require.Equal(t, instructions[0], storeInstruction)
 
 	expectedOutput := fmt.Sprintf(`In the module that calls store.star
 In the store files instruction
-{{kurtosis:%v-4:39.artifact_uuid}}
-`, storeFileDefinitionPath)
+%v
+`, testArtifactUuid)
 	require.Equal(t, expectedOutput, string(scriptOutput))
 }
 
@@ -1184,14 +1195,16 @@ The service example-datastore-server has been removed
 
 func TestStartosisInterpreter_UploadGetsInterpretedCorrectly(t *testing.T) {
 	filePath := "github.com/kurtosis/module/lib/lib.star"
+	artifactUuid, err := enclave_data_directory.NewFilesArtifactUUID()
+	require.Nil(t, err)
 	moduleContentProvider := mock_module_content_provider.NewMockModuleContentProvider()
 	defer moduleContentProvider.RemoveAll()
-	err := moduleContentProvider.AddFileContent(filePath, "fooBar")
+	err = moduleContentProvider.AddFileContent(filePath, "fooBar")
 	require.Nil(t, err)
 	filePathOnDisk, err := moduleContentProvider.GetOnDiskAbsoluteFilePath(filePath)
 	require.Nil(t, err)
 	interpreter := NewStartosisInterpreter(testServiceNetwork, moduleContentProvider)
-	script := `upload_files("` + filePath + `")
+	script := `upload_files("` + filePath + `","` + string(artifactUuid) + `")
 `
 	scriptOutput, interpretationError, instructions := interpreter.Interpret(context.Background(), ModuleIdPlaceholderForStandaloneScripts, script, EmptyInputArgs)
 	require.Nil(t, interpretationError)
@@ -1200,7 +1213,7 @@ func TestStartosisInterpreter_UploadGetsInterpretedCorrectly(t *testing.T) {
 
 	expectedUploadInstruction := upload_files.NewUploadFilesInstruction(
 		*kurtosis_instruction.NewInstructionPosition(1, 13, starlarkFilenamePlaceholderAsNotUsed),
-		testServiceNetwork, moduleContentProvider, filePath, filePathOnDisk,
+		testServiceNetwork, moduleContentProvider, filePath, filePathOnDisk, artifactUuid,
 	)
 
 	require.Equal(t, expectedUploadInstruction, instructions[0])
@@ -1208,16 +1221,18 @@ func TestStartosisInterpreter_UploadGetsInterpretedCorrectly(t *testing.T) {
 
 func createSimpleAddServiceInstruction(t *testing.T, serviceId service.ServiceID, imageName string, portNumber uint32, lineNumber int32, colNumber int32, fileName string, entryPointArgs []string, cmdArgs []string, envVars map[string]string, privateIPAddressPlaceholder string) *add_service.AddServiceInstruction {
 	serviceConfigStringDict := starlark.StringDict{}
-	serviceConfigStringDict["container_image_name"] = starlark.String(imageName)
+	serviceConfigStringDict["image"] = starlark.String(imageName)
 
-	usedPortDict := starlark.NewDict(1)
-	require.Nil(t, usedPortDict.SetKey(
-		starlark.String("grpc"),
-		starlarkstruct.FromStringDict(starlarkstruct.Default, starlark.StringDict{
-			"number":   starlark.MakeInt(int(portNumber)),
-			"protocol": starlark.String("TCP"),
-		})))
-	serviceConfigStringDict["used_ports"] = usedPortDict
+	if portNumber != 0 {
+		usedPortDict := starlark.NewDict(1)
+		require.Nil(t, usedPortDict.SetKey(
+			starlark.String("grpc"),
+			starlarkstruct.FromStringDict(starlarkstruct.Default, starlark.StringDict{
+				"number":   starlark.MakeInt(int(portNumber)),
+				"protocol": starlark.String("TCP"),
+			})))
+		serviceConfigStringDict["ports"] = usedPortDict
+	}
 
 	if entryPointArgs != nil {
 		entryPointArgsValues := make([]starlark.Value, 0)
@@ -1253,14 +1268,19 @@ func createSimpleAddServiceInstruction(t *testing.T, serviceId service.ServiceID
 
 	serviceConfigBuilder := services.NewServiceConfigBuilder(
 		imageName,
-	).WithPrivatePorts(
-		map[string]*kurtosis_core_rpc_api_bindings.Port{
-			"grpc": {
-				Number:   portNumber,
-				Protocol: kurtosis_core_rpc_api_bindings.Port_TCP,
-			},
-		},
 	)
+
+	if portNumber != 0 {
+		serviceConfigBuilder.WithPrivatePorts(
+			map[string]*kurtosis_core_rpc_api_bindings.Port{
+				"grpc": {
+					Number:   portNumber,
+					Protocol: kurtosis_core_rpc_api_bindings.Port_TCP,
+				},
+			},
+		)
+	}
+
 	if entryPointArgs != nil {
 		serviceConfigBuilder.WithEntryPointArgs(entryPointArgs)
 	}
@@ -1281,8 +1301,8 @@ func createSimpleAddServiceInstruction(t *testing.T, serviceId service.ServiceID
 		serviceId,
 		serviceConfigBuilder.Build(),
 		starlark.StringDict{
-			"service_id":     starlark.String(serviceId),
-			"service_config": serviceConfigStruct,
+			"service_id": starlark.String(serviceId),
+			"config":     serviceConfigStruct,
 		},
 	)
 }
@@ -1296,4 +1316,35 @@ func TestStartosisInterpreter_NoPanicIfUploadIsPassedAPathNotOnDisk(t *testing.T
 `
 	_, interpretationError, _ := interpreter.Interpret(context.Background(), ModuleIdPlaceholderForStandaloneScripts, script, EmptyInputArgs)
 	require.NotNil(t, interpretationError)
+}
+
+func TestStartosisInterpreter_NoPortsIsOkayForAddServiceInstruction(t *testing.T) {
+	moduleContentProvider := mock_module_content_provider.NewMockModuleContentProvider()
+	defer moduleContentProvider.RemoveAll()
+	interpreter := NewStartosisInterpreter(testServiceNetwork, moduleContentProvider)
+	script := `
+print("Starting Startosis script!")
+
+service_id = "example-datastore-server"
+print("Adding service " + service_id)
+
+config = struct(
+	image = "` + testContainerImageName + `",
+)
+datastore_service = add_service(service_id = service_id, config = config)
+print("The datastore service ip address is " + datastore_service.ip_address)
+`
+
+	scriptOutput, interpretationError, instructions := interpreter.Interpret(context.Background(), ModuleIdPlaceholderForStandaloneScripts, script, EmptyInputArgs)
+	require.Nil(t, interpretationError)
+	require.Equal(t, 1, len(instructions))
+
+	addServiceInstruction := createSimpleAddServiceInstruction(t, "example-datastore-server", testContainerImageName, 0, 10, 32, starlarkFilenamePlaceholderAsNotUsed, defaultEntryPointArgs, defaultCmdArgs, defaultEnvVars, defaultPrivateIPAddressPlaceholder)
+	require.Equal(t, instructions[0], addServiceInstruction)
+
+	expectedOutput := `Starting Startosis script!
+Adding service example-datastore-server
+The datastore service ip address is {{kurtosis:example-datastore-server.ip_address}}
+`
+	require.Equal(t, expectedOutput, string(scriptOutput))
 }